issue #112 refacotor: common carousel refactoring
