+++ src/components/common/Carousel/Carousel.tsx
@@ -1,4 +1,4 @@
-import { createContext, useMemo } from "react";
+import { createContext, useMemo, useState, useRef } from "react";
 import type { PropsWithChildren } from "react";
 
 import LeftArrowIcon from "@/assets/svg/ic-left-arrow.svg?react";
@@ -7,8 +7,10 @@ import RightArrowIcon from "@/assets/svg/ic-right-arrow.svg?react";
 import { Box } from "@/components/common";
 import CarouselItem from "@/components/common/Carousel/CarouselItem";
 import Dots from "@/components/common/Carousel/Dots";
+import Gallery from "@/components/common/Carousel/Gallery/Gallery";
 
 import useCarousel from "@/hooks/useCarousel";
+import useClickOutSide from "@/hooks/useClickOutSide";
 
 import {
 	carouselStyle,
@@ -25,7 +27,9 @@ export interface CarouselProps extends PropsWithChildren {
 	length: number;
 	showArrows?: boolean;
 	showDots?: boolean;
-	children?: JSX.Element | JSX.Element[];
+	updateMediaList?: string[];
+	setUpdateMediaList?: React.Dispatch<React.SetStateAction<string[]>>;
+	hasGallery?: boolean;
 }
 
 export const CarouselContext = createContext<{
@@ -43,6 +47,9 @@ const Carousel = ({
 	showArrows = true,
 	showDots = true,
 	children,
+	updateMediaList,
+	setUpdateMediaList,
+	hasGallery,
 }: CarouselProps) => {
 	const {
 		mediaIndex,
@@ -76,9 +83,33 @@ const Carousel = ({
 		],
 	);
 
+	const [isGalleryOpen, setIsGalleryOpen] = useState(false);
+	const [, setMediaCurrentIndex] = useState(0);
+
+	const galleryRef = useRef<HTMLDivElement>(null);
+
+	useClickOutSide(galleryRef, () => setIsGalleryOpen(false));
+
+	const handleGalleryOpen = () => {
+		setIsGalleryOpen((prev) => !prev);
+	};
+
 	return (
 		<CarouselContext.Provider value={context}>
 			<div css={carouselStyle(width, height, borderRadius)} ref={carouselBoxRef}>
+				{hasGallery && (
+					<Gallery
+						isGalleryOpen={isGalleryOpen}
+						handleGalleryOpen={handleGalleryOpen}
+						galleryRef={galleryRef}
+						mediaCurrentIndex={mediaIndex}
+						setMediaCurrentIndex={setMediaCurrentIndex}
+						updatedMediaList={updateMediaList}
+						setUpdateMediaList={setUpdateMediaList}
+						handleMoveImage={handleMoveImage}
+					/>
+				)}
+
 				{showArrows && length !== 1 && (
 					<div css={buttonBoxStyle}>
 						<button type="button" css={leftButtonStyle} onClick={handleClickLeft}>
