+++ src/components/common/Carousel/Gallery/Gallery.style.ts
