+++ src/components/common/Carousel/Gallery/GallerySlider/GallerySlider.style.ts
